Improved lesson progress list. Interaction model course registration. Tests.
Each course page's lesson list now includes all the lessons in the course,
instead of just the previously visited lessons.

Also, lesson visit details are only recorded where the user is registered on
the course. This is implemented at model level and requires views integration.

Finally, tests in place to check that para 2 above is implemented.